[lib] first try of `QuantumToolboxUtils`
